Using pluralforms requires an int, not a QString that (hopefully) represents the number. Since I don't know the code, this fix is a guess.
svn path=/trunk/kdebase/ksysguard/; revision=149856

+++ gui/ksysguard.cc
@@ -460,7 +460,10 @@ TopLevel::answerReceived(int id, const QString& answer)
 	case 0:
 		// yes, I know there is never 1 process, but that's the way
 		// singular vs. plural works :/
-		s = i18n("1 Process", "%n Processes", answer);
+		//
+		// To use pluralForms, though, you need to convert to
+		// an integer, not use the QString straight.
+		s = i18n("1 Process", "%n Processes", answer.toInt());
 		statusbar->changeItem(s, 0);
 		break;
 	case 1:
